backend/community_analytics/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
google-cloud-firestore==2.16.0
transformers==4.43.4
torch>=2.2.0
sentence-transformers==3.0.1
bertopic==0.16.3
hdbscan==0.8.38.post1
umap-learn==0.5.6
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0
